+++ cmd/mount.go
@@ -70,24 +70,26 @@ type LogOptions struct {
 }
 
 type mountOptions struct {
-	MountPath  string
-	ConfigFile string
+	MountPath      string
+	inputMountPath string
+	ConfigFile     string
 
-	Logging           LogOptions     `config:"logging"`
-	Components        []string       `config:"components"`
-	Foreground        bool           `config:"foreground"`
-	NonEmpty          bool           `config:"nonempty"`
-	DefaultWorkingDir string         `config:"default-working-dir"`
-	CPUProfile        string         `config:"cpu-profile"`
-	MemProfile        string         `config:"mem-profile"`
-	PassPhrase        string         `config:"passphrase"`
-	SecureConfig      bool           `config:"secure-config"`
-	DynamicProfiler   bool           `config:"dynamic-profile"`
-	ProfilerPort      int            `config:"profiler-port"`
-	ProfilerIP        string         `config:"profiler-ip"`
-	MonitorOpt        monitorOptions `config:"health_monitor"`
-	WaitForMount      time.Duration  `config:"wait-for-mount"`
-	LazyWrite         bool           `config:"lazy-write"`
+	Logging            LogOptions     `config:"logging"`
+	Components         []string       `config:"components"`
+	Foreground         bool           `config:"foreground"`
+	NonEmpty           bool           `config:"nonempty"`
+	DefaultWorkingDir  string         `config:"default-working-dir"`
+	CPUProfile         string         `config:"cpu-profile"`
+	MemProfile         string         `config:"mem-profile"`
+	PassPhrase         string         `config:"passphrase"`
+	SecureConfig       bool           `config:"secure-config"`
+	DynamicProfiler    bool           `config:"dynamic-profile"`
+	ProfilerPort       int            `config:"profiler-port"`
+	ProfilerIP         string         `config:"profiler-ip"`
+	MonitorOpt         monitorOptions `config:"health_monitor"`
+	WaitForMount       time.Duration  `config:"wait-for-mount"`
+	LazyWrite          bool           `config:"lazy-write"`
+	disableKernelCache bool           `config:"disable-kernel-cache"`
 
 	// v1 support
 	Streaming         bool     `config:"streaming"`
@@ -110,28 +112,15 @@ func (opt *mountOptions) validate(skipNonEmptyMount bool) error {
 	} else if common.IsDirectoryMounted(opt.MountPath) {
 		// Try to cleanup the stale mount
 		log.Info("Mount::validate : Mount directory is already mounted, trying to cleanup")
-		active, err := common.IsMountActive(opt.MountPath)
+		active, err := common.IsMountActive(opt.inputMountPath)
 		if active || err != nil {
 			// Previous mount is still active so we need to fail this mount
 			return fmt.Errorf("directory is already mounted")
 		} else {
 			// Previous mount is in stale state so lets cleanup the state
 			log.Info("Mount::validate : Cleaning up stale mount")
-			if err = unmountBlobfuse2(opt.MountPath); err != nil {
+			if err = unmountBlobfuse2(opt.MountPath, true, true); err != nil {
 				return fmt.Errorf("directory is already mounted, unmount manually before remount [%v]", err.Error())
-			}
-
-			// Clean up the file-cache temp directory if any
-			var tempCachePath string
-			_ = config.UnmarshalKey("file_cache.path", &tempCachePath)
-
-			var cleanupOnStart bool
-			_ = config.UnmarshalKey("file_cache.cleanup-on-start", &cleanupOnStart)
-
-			if tempCachePath != "" && cleanupOnStart {
-				if err = common.TempCacheCleanup(tempCachePath); err != nil {
-					return fmt.Errorf("failed to cleanup file cache [%s]", err.Error())
-				}
 			}
 		}
 	} else if !skipNonEmptyMount && !common.IsDirectoryEmpty(opt.MountPath) {
@@ -264,10 +253,12 @@ var mountCmd = &cobra.Command{
 	Args:              cobra.ExactArgs(1),
 	FlagErrorHandling: cobra.ExitOnError,
 	RunE: func(_ *cobra.Command, args []string) error {
+		options.inputMountPath = args[0]
 		options.MountPath = common.ExpandPath(args[0])
 		common.MountPath = options.MountPath
 
 		configFileExists := true
+		directIO := false
 
 		if options.ConfigFile == "" {
 			// Config file is not set in cli parameters
@@ -388,10 +379,24 @@ var mountCmd = &cobra.Command{
 				} else if v == "direct_io" || v == "direct_io=true" {
 					config.Set("lfuse.direct-io", "true")
 					config.Set("direct-io", "true")
+					directIO = true
 				} else {
 					return errors.New(common.FuseAllowedFlags)
 				}
 			}
+		}
+
+		// Check if direct-io is enabled in the config file.
+		if !directIO {
+			_ = config.UnmarshalKey("libfuse.direct-io", &directIO)
+			if directIO {
+				config.Set("direct-io", "true")
+			}
+		}
+
+		if config.IsSet("disable-kernel-cache") && directIO {
+			// Both flag shall not be enable together
+			return fmt.Errorf("direct-io and disable-kernel-cache cannot be enabled together")
 		}
 
 		if !config.IsSet("logging.file-path") {
@@ -461,10 +466,8 @@ var mountCmd = &cobra.Command{
 		log.Crit("Logging level set to : %s", logLevel.String())
 		log.Debug("Mount allowed on nonempty path : %v", options.NonEmpty)
 
-		directIO := false
-		_ = config.UnmarshalKey("direct-io", &directIO)
 		if directIO {
-			// Directio is enabled, so remove the attr-cache from the pipeline
+			// Direct IO is enabled, so remove the attr-cache from the pipeline
 			for i, name := range options.Components {
 				if name == "attr_cache" {
 					options.Components = append(options.Components[:i], options.Components[i+1:]...)
@@ -474,13 +477,20 @@ var mountCmd = &cobra.Command{
 			}
 		}
 
+		// Clean up any cache directory if cleanup-on-start is set from the cli parameter or specified in parameter in
+		// config file for a specific component for file-cache, block-cache, xload.
+		err = options.tempCacheCleanup()
+		if err != nil {
+			return err
+		}
+
+		common.ForegroundMount = options.Foreground
+
 		pipeline, err = internal.NewPipeline(options.Components, !daemon.WasReborn())
 		if err != nil {
 			log.Err("mount : failed to initialize new pipeline [%v]", err)
 			return Destroy(fmt.Sprintf("failed to initialize new pipeline [%s]", err.Error()))
 		}
-
-		common.ForegroundMount = options.Foreground
 
 		log.Info("mount: Mounting blobfuse2 on %s", options.MountPath)
 		if !options.Foreground {
@@ -512,10 +522,18 @@ var mountCmd = &cobra.Command{
 				}()
 			}
 
+		retry:
+			// If the .pid file is locked and there no blobfuse process owning it then we need to try
+			// a cleanup of the .pid file. If cleanup goes through then retry the daemonization.
 			child, err := dmnCtx.Reborn()
 			if err != nil {
-				log.Err("mount : failed to daemonize application [%v]", err)
-				return Destroy(fmt.Sprintf("failed to daemonize application [%s]", err.Error()))
+				log.Err("mount : failed to daemonize application [%s], trying auto cleanup", err.Error())
+				rmErr := os.Remove(pidFileName)
+				if rmErr != nil {
+					log.Err("mount : auto cleanup failed [%v]", rmErr.Error())
+					return Destroy(fmt.Sprintf("failed to daemonize application [%s]", err.Error()))
+				}
+				goto retry
 			}
 
 			log.Debug("mount: foreground disabled, child = %v", daemon.WasReborn())
@@ -671,6 +689,51 @@ func startMonitor(pid int) {
 	}
 }
 
+// cleanupCachePath is a helper function to clean up cache directory for a component that is present in the pipeline.
+// componentName: the name of the component (e.g., "file_cache", "block_cache")
+func (opt *mountOptions) tempCacheCleanup() error {
+	// Check for global cleanup-on-start flag from cli.
+	var cleanupOnStart bool
+	_ = config.UnmarshalKey("cleanup-on-start", &cleanupOnStart)
+
+	components := []string{"file_cache", "block_cache", "xload"}
+
+	for _, component := range components {
+		if common.ComponentInPipeline(options.Components, component) {
+			err := cleanupCachePath(component, cleanupOnStart)
+			if err != nil {
+				return fmt.Errorf("failed to clean up  cache for %s: %v", component, err)
+			}
+		}
+	}
+
+	return nil
+}
+
+func cleanupCachePath(componentName string, globalCleanupFlag bool) error {
+	// Get the path for the component
+	var cachePath string
+	_ = config.UnmarshalKey(componentName+".path", &cachePath)
+
+	if cachePath == "" {
+		// No path configured for this component
+		return nil
+	}
+
+	// Check for component-specific cleanup flag
+	var componentCleanupFlag bool
+	_ = config.UnmarshalKey(componentName+".cleanup-on-start", &componentCleanupFlag)
+
+	// Clean up if either global or component-specific flag is set
+	if globalCleanupFlag || componentCleanupFlag {
+		if err := common.TempCacheCleanup(cachePath); err != nil {
+			return fmt.Errorf("failed to cleanup temp cache path: %s for %s component: %v", cachePath, componentName, err)
+		}
+	}
+
+	return nil
+}
+
 func sigusrHandler(pipeline *internal.Pipeline, ctx context.Context) daemon.SignalHandlerFunc {
 	return func(sig os.Signal) error {
 		log.Crit("Mount::sigusrHandler : Signal %d received", sig)
@@ -752,6 +815,10 @@ func init() {
 		return []string{"silent", "base", "syslog"}, cobra.ShellCompDirectiveNoFileComp
 	})
 
+	// Add a generic cleanup-on-start flag that applies to all cache components
+	mountCmd.PersistentFlags().Bool("cleanup-on-start", false, "Clear cache directory on startup if not empty for file_cache, block_cache, xload components.")
+	config.BindPFlag("cleanup-on-start", mountCmd.PersistentFlags().Lookup("cleanup-on-start"))
+
 	mountCmd.PersistentFlags().String("log-level", "LOG_WARNING",
 		"Enables logs written to syslog. Set to LOG_WARNING by default. Allowed values are LOG_OFF|LOG_CRIT|LOG_ERR|LOG_WARNING|LOG_INFO|LOG_DEBUG")
 	config.BindPFlag("logging.level", mountCmd.PersistentFlags().Lookup("log-level"))
@@ -808,6 +875,9 @@ func init() {
 	mountCmd.PersistentFlags().ShorthandLookup("o").Hidden = true
 
 	mountCmd.PersistentFlags().DurationVar(&options.WaitForMount, "wait-for-mount", 5*time.Second, "Let parent process wait for given timeout before exit")
+
+	mountCmd.PersistentFlags().Bool("disable-kernel-cache", false, "Disable kerneel cache, but keep blobfuse cache. Default value false.")
+	config.BindPFlag("disable-kernel-cache", mountCmd.PersistentFlags().Lookup("disable-kernel-cache"))
 
 	config.AttachToFlagSet(mountCmd.PersistentFlags())
 	config.AttachFlagCompletions(mountCmd)