+++ cmd/mount_test.go
@@ -38,9 +38,11 @@ import (
 	"io"
 	"os"
 	"path/filepath"
+	"strings"
 	"testing"
 
 	"github.com/Azure/azure-storage-fuse/v2/common"
+	"github.com/Azure/azure-storage-fuse/v2/common/config"
 	"github.com/Azure/azure-storage-fuse/v2/common/log"
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/assert"
@@ -96,6 +98,22 @@ components:
   - azstorage
 `
 
+var configDirectIOTest string = `
+libfuse:
+  direct-io: true
+azstorage:
+  account-name: myAccountName
+  account-key: myAccountKey
+  mode: key
+  endpoint: myEndpoint
+  container: myContainer
+  max-retries: 1
+components:
+  - libfuse
+  - attr_cache
+  - azstorage
+`
+
 var confFileMntTest, confFilePriorityTest string
 
 type mountTestSuite struct {
@@ -109,6 +127,14 @@ func (suite *mountTestSuite) SetupTest() {
 	err := log.SetDefaultLogger("silent", common.LogConfig{Level: common.ELogLevel.LOG_DEBUG()})
 	if err != nil {
 		panic("Unable to set silent logger as default.")
+	}
+}
+
+func (suite *mountTestSuite) SetupSuite() {
+	out, err := executeCommandC(rootCmd, "mount")
+	strings.Contains(out, "accepts 1 arg(s), received 0")
+	if err == nil {
+		panic("Unable to parse the empty flags to mount command")
 	}
 }
 
@@ -238,6 +264,34 @@ func (suite *mountTestSuite) TestComponentPrioritySetWrong() {
 	suite.assert.Contains(op, "component libfuse is out of order")
 }
 
+func (suite *mountTestSuite) TestDirectIOInConfig() {
+	defer suite.cleanupTest()
+
+	mntDir, err := os.MkdirTemp("", "mntdir")
+	suite.assert.NoError(err)
+	defer os.RemoveAll(mntDir)
+
+	confFile, err := os.CreateTemp("", "conf*.yaml")
+	suite.assert.NoError(err)
+	confFileName := confFile.Name()
+	defer os.Remove(confFileName)
+
+	_, err = confFile.WriteString(configDirectIOTest)
+	suite.assert.NoError(err)
+	confFile.Close()
+
+	op, err := executeCommandC(rootCmd, "mount", mntDir, fmt.Sprintf("--config-file=%s", confFileName))
+	suite.assert.Error(err)
+	suite.assert.Contains(op, "failed to initialize new pipeline")
+
+	directIO := false
+	err = config.UnmarshalKey("direct-io", &directIO)
+	suite.assert.NoError(err)
+	suite.assert.True(directIO)
+
+	suite.assert.NotContains(options.Components, "attr_cache")
+}
+
 func (suite *mountTestSuite) TestDefaultConfigFile() {
 	defer suite.cleanupTest()
 
@@ -313,6 +367,21 @@ func (suite *mountTestSuite) TestStreamAttrCacheOptionsV1() {
 		"--streaming", "--use-attr-cache", "--invalidate-on-sync", "--pre-mount-validate", "--basic-remount-check", "-o", "direct_io")
 	suite.assert.NotNil(err)
 	suite.assert.Contains(op, "failed to initialize new pipeline")
+}
+
+func (suite *mountTestSuite) TestDirectIODisableKernelCacheCombo() {
+	defer suite.cleanupTest()
+
+	mntDir, err := os.MkdirTemp("", "mntdir")
+	suite.assert.Nil(err)
+	defer os.RemoveAll(mntDir)
+
+	tempLogDir := "/tmp/templogs_" + randomString(6)
+	defer os.RemoveAll(tempLogDir)
+
+	op, err := executeCommandC(rootCmd, "mount", mntDir, "-o", "direct_io", "--disable-kernel-cache")
+	suite.assert.NotNil(err)
+	suite.assert.Contains(op, "direct-io and disable-kernel-cache cannot be enabled together")
 }
 
 // mount failure test where a libfuse option is incorrect
@@ -493,6 +562,90 @@ func (suite *mountTestSuite) TestMountOptionVaildate() {
 	suite.assert.Contains(opts.Logging.LogFilePath, opts.DefaultWorkingDir)
 	suite.assert.Equal(common.DefaultWorkDir, opts.DefaultWorkingDir)
 	suite.assert.Equal(common.DefaultLogFilePath, opts.Logging.LogFilePath)
+}
+
+func (suite *mountTestSuite) TestCleanUpOnStartFlag() {
+	defer suite.cleanupTest()
+	// Create a test directory
+	testDir := filepath.Join(os.TempDir(), "cleanup_test")
+	os.RemoveAll(testDir)
+	os.MkdirAll(testDir, 0755)
+
+	defer func() {
+		os.RemoveAll(testDir)
+	}()
+
+	testPath := filepath.Join(testDir, "dir1")
+	testPath2 := filepath.Join(testDir, "dir2")
+	testPath3 := filepath.Join(testDir, "dir3")
+	os.MkdirAll(testPath, 0755)
+	os.MkdirAll(testPath2, 0755)
+	os.MkdirAll(testPath3, 0755)
+
+	createFilesInCacheDirs := func() {
+		// Create some test files
+		testFile := filepath.Join(testPath, "testfile")
+		err := os.WriteFile(testFile, []byte("test"), 0644)
+		suite.assert.Nil(err)
+
+		testFile2 := filepath.Join(testPath2, "testfile")
+		err = os.WriteFile(testFile2, []byte("test"), 0644)
+		suite.assert.Nil(err)
+
+		testFile3 := filepath.Join(testPath3, "testfile")
+		err = os.WriteFile(testFile3, []byte("test"), 0644)
+		suite.assert.Nil(err)
+	}
+
+	comps := []string{"file_cache", "block_cache", "xload"}
+	var cachedirs []string
+	cachedirs = append(cachedirs, testPath, testPath2, testPath3)
+
+	// Set the path for the cache directory.
+	config.Set(comps[0]+".path", testPath)
+	config.Set(comps[1]+".path", testPath2)
+	config.Set(comps[2]+".path", testPath3)
+
+	for i, comp := range comps {
+		// Set up test component
+		testComponent := comp
+		options.Components = []string{testComponent}
+
+		// Test case 1: Global flag true, component flag false
+		createFilesInCacheDirs()
+		config.Set("cleanup-on-start", "true")
+		config.Set(testComponent+".cleanup-on-start", "false")
+
+		err := options.tempCacheCleanup()
+		suite.assert.Nil(err)
+		suite.assert.True(common.IsDirectoryEmpty(cachedirs[i]))
+		// This should not delete the other cache dirs of other components.
+		suite.assert.False(common.IsDirectoryEmpty(cachedirs[(i+1)%3]))
+		suite.assert.False(common.IsDirectoryEmpty(cachedirs[(i+2)%3]))
+
+		// Test case 2: Global flag false, component flag true
+		createFilesInCacheDirs()
+		config.Set(testComponent+".cleanup-on-start", "true")
+		config.Set("cleanup-on-start", "false")
+
+		err = options.tempCacheCleanup()
+		suite.assert.Nil(err)
+		suite.assert.True(common.IsDirectoryEmpty(cachedirs[i]))
+		// This should not delete the other cache dirs of other components.
+		suite.assert.False(common.IsDirectoryEmpty(cachedirs[(i+1)%3]))
+		suite.assert.False(common.IsDirectoryEmpty(cachedirs[(i+2)%3]))
+
+		// Test case 3: Both flags false
+		createFilesInCacheDirs()
+		config.Set(testComponent+".cleanup-on-start", "false")
+		config.Set("cleanup-on-start", "false")
+		err = options.tempCacheCleanup()
+		suite.assert.Nil(err)
+		suite.assert.False(common.IsDirectoryEmpty(cachedirs[i]))
+		// This should not delete the other cache dirs of other components.
+		suite.assert.False(common.IsDirectoryEmpty(cachedirs[(i+1)%3]))
+		suite.assert.False(common.IsDirectoryEmpty(cachedirs[(i+2)%3]))
+	}
 }
 
 func TestMountCommand(t *testing.T) {