+++ cmd/unmount.go
@@ -53,7 +53,8 @@ var unmountCmd = &cobra.Command{
 	SuggestFor:        []string{"unmount", "unmnt"},
 	Args:              cobra.ExactArgs(1),
 	FlagErrorHandling: cobra.ExitOnError,
-	RunE: func(_ *cobra.Command, args []string) error {
+	RunE: func(cmd *cobra.Command, args []string) error {
+		lazy, _ := cmd.Flags().GetBool("lazy")
 		if strings.Contains(args[0], "*") {
 			mntPathPrefix := args[0]
 
@@ -61,14 +62,14 @@ var unmountCmd = &cobra.Command{
 			for _, mntPath := range lstMnt {
 				match, _ := regexp.MatchString(mntPathPrefix, mntPath)
 				if match {
-					err := unmountBlobfuse2(mntPath)
+					err := unmountBlobfuse2(mntPath, lazy, false)
 					if err != nil {
 						return fmt.Errorf("failed to unmount %s [%s]", mntPath, err.Error())
 					}
 				}
 			}
 		} else {
-			err := unmountBlobfuse2(args[0])
+			err := unmountBlobfuse2(args[0], lazy, false)
 			if err != nil {
 				return err
 			}
@@ -86,18 +87,26 @@ var unmountCmd = &cobra.Command{
 }
 
 // Attempts to unmount the directory and returns true if the operation succeeded
-func unmountBlobfuse2(mntPath string) error {
+func unmountBlobfuse2(mntPath string, lazy bool, silent bool) error {
 	unmountCmd := []string{"fusermount3", "fusermount"}
 
 	var errb bytes.Buffer
 	var err error
 	for _, umntCmd := range unmountCmd {
-		cliOut := exec.Command(umntCmd, "-u", mntPath)
+		var args []string
+		if lazy {
+			args = append(args, "-z")
+		}
+		args = append(args, "-u", mntPath)
+		cliOut := exec.Command(umntCmd, args...)
 		cliOut.Stderr = &errb
 		_, err = cliOut.Output()
 
 		if err == nil {
-			fmt.Println("Successfully unmounted", mntPath)
+			log.Info("unmountBlobfuse2 : successfully unmounted %s", mntPath)
+			if !silent {
+				fmt.Println("Successfully unmounted", mntPath)
+			}
 			return nil
 		}
 
@@ -113,4 +122,6 @@ func unmountBlobfuse2(mntPath string) error {
 func init() {
 	rootCmd.AddCommand(unmountCmd)
 	unmountCmd.AddCommand(umntAllCmd)
+
+	unmountCmd.PersistentFlags().BoolP("lazy", "z", false, "Use lazy unmount")
 }