+++ common/config/config_parser.go
@@ -46,7 +46,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/fsnotify/fsnotify"
-	"github.com/mitchellh/mapstructure"
+	"github.com/go-viper/mapstructure/v2"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
 )
@@ -146,6 +146,7 @@ func DecryptConfigFile(fileName string, passphrase string) error {
 		return fmt.Errorf("Failed to decrypt config file [%s]", err.Error())
 	}
 
+	viper.SetConfigType("yaml")
 	err = loadConfigFromBufferToViper(plainText)
 	if err != nil {
 		return fmt.Errorf("Failed to load decrypted config file [%s]", err.Error())