+++ component/xload/splitter.go
@@ -107,9 +107,9 @@ func (ds *downloadSplitter) Init() {
 	}
 }
 
-func (ds *downloadSplitter) Start() {
+func (ds *downloadSplitter) Start(ctx context.Context) {
 	log.Debug("downloadSplitter::Start : start download splitter for %s", ds.path)
-	ds.GetThreadPool().Start()
+	ds.GetThreadPool().Start(ctx)
 }
 
 func (ds *downloadSplitter) Stop() {
@@ -117,7 +117,7 @@ func (ds *downloadSplitter) Stop() {
 	if ds.GetThreadPool() != nil {
 		ds.GetThreadPool().Stop()
 	}
-	ds.GetNext().Stop()
+	log.Debug("downloadSplitter::Stop : stop successful")
 }
 
 // download data in chunks and then write to the local file
@@ -202,23 +202,39 @@ func (ds *downloadSplitter) Process(item *WorkItem) (int, error) {
 		defer wg.Done()
 
 		for i := 0; i < int(numBlocks); i++ {
-			respSplitItem := <-responseChannel
-			if respSplitItem.Err != nil {
-				log.Err("downloadSplitter::Process : Failed to download data for file %s", item.Path)
+			select {
+			case <-ds.GetThreadPool().ctx.Done(): // check if the thread pool is closed
 				operationSuccess = false
-				cancel() // cancel the context to stop download of other chunks
-			} else {
-				_, err := item.FileHandle.WriteAt(respSplitItem.Block.Data[:respSplitItem.DataLen], respSplitItem.Block.Offset)
-				if err != nil {
-					log.Err("downloadSplitter::Process : Failed to write data to file %s", item.Path)
+				cancel()
+				return
+			case respSplitItem := <-responseChannel:
+				if respSplitItem.Err != nil {
+					log.Err("downloadSplitter::Process : Failed to download data for file %s", item.Path)
 					operationSuccess = false
 					cancel() // cancel the context to stop download of other chunks
-				}
-			}
+				} else {
+					_, err := item.FileHandle.WriteAt(respSplitItem.Block.Data[:respSplitItem.DataLen], respSplitItem.Block.Offset)
+					if err != nil {
+						log.Err("downloadSplitter::Process : Failed to write data to file %s [%s]", item.Path, err.Error())
+						operationSuccess = false
+						cancel() // cancel the context to stop download of other chunks
+					}
 
-			if respSplitItem.Block != nil {
-				// log.Debug("downloadSplitter::process : Download successful %s index %d offset %v", item.path, respSplitItem.block.index, respSplitItem.block.offset)
-				ds.blockPool.Release(respSplitItem.Block)
+					// send the download status to stats manager
+					ds.GetStatsManager().AddStats(&StatsItem{
+						Component:        SPLITTER,
+						Name:             item.Path,
+						Success:          false,
+						Download:         false,
+						DiskIO:           true,
+						BytesTransferred: respSplitItem.DataLen,
+					})
+				}
+
+				if respSplitItem.Block != nil {
+					// log.Debug("downloadSplitter::process : Download successful %s index %d offset %v", item.path, respSplitItem.block.index, respSplitItem.block.offset)
+					ds.blockPool.Release(respSplitItem.Block)
+				}
 			}
 		}
 	}()
@@ -244,7 +260,11 @@ func (ds *downloadSplitter) Process(item *WorkItem) (int, error) {
 				Ctx:             ctx,
 			}
 			// log.Debug("downloadSplitter::Process : Scheduling download for %s offset %v", item.Path, offset)
-			ds.GetNext().Schedule(splitItem)
+			err := ds.GetNext().Schedule(splitItem)
+			if err != nil {
+				log.Err("downloadSplitter::Process : Failed to schedule download for %s [%s]", item.Path, err.Error())
+				responseChannel <- &WorkItem{Err: fmt.Errorf("failed to schedule download for %s [%s]", item.Path, err.Error())}
+			}
 		}
 
 		offset += int64(ds.blockPool.GetBlockSize())
@@ -252,19 +272,21 @@ func (ds *downloadSplitter) Process(item *WorkItem) (int, error) {
 
 	wg.Wait()
 
-	// update the last modified time
-	// TODO:: xload : verify if the lmt is updated correctly
-	err = os.Chtimes(localPath, item.Atime, item.Mtime)
-	if err != nil {
-		log.Err("downloadSplitter::Process : Failed to change times of file %s [%s]", item.Path, err.Error())
-	}
-
-	if ds.validateMD5 && operationSuccess {
-		err = ds.checkConsistency(item)
+	if operationSuccess {
+		// update the last modified time
+		// TODO:: xload : verify if the lmt is updated correctly
+		err = os.Chtimes(localPath, item.Atime, item.Mtime)
 		if err != nil {
-			// TODO:: xload : retry if md5 validation fails
-			log.Err("downloadSplitter::Process : unable to validate md5 for %s [%s]", item.Path, err.Error())
-			operationSuccess = false
+			log.Err("downloadSplitter::Process : Failed to change times of file %s [%s]", item.Path, err.Error())
+		}
+
+		if ds.validateMD5 {
+			err = ds.checkConsistency(item)
+			if err != nil {
+				// TODO:: xload : retry if md5 validation fails
+				log.Err("downloadSplitter::Process : unable to validate md5 for %s [%s]", item.Path, err.Error())
+				operationSuccess = false
+			}
 		}
 	}