+++ internal/component.go
@@ -84,7 +84,6 @@ type Component interface {
 	CreateDir(CreateDirOptions) error
 	DeleteDir(DeleteDirOptions) error
 	IsDirEmpty(IsDirEmptyOptions) bool
-	DeleteEmptyDirs(DeleteDirOptions) (bool, error)
 
 	OpenDir(OpenDirOptions) error
 	//ReadDir: implementation expectations